app/requirements.txt
fastapi
redis
uvicorn
passlib
python-jose
bcrypt
pymysql
requests
asgi-lifespan
python-dateutil
httpx
sqlmodel